backend/requirements.txt
asgiref==3.8.1
astroid==3.2.4
attrs==23.2.0
black==24.4.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==7.6.0
cryptography==43.0.0
defusedxml==0.8.0rc2
dill==0.3.8
Django==5.0.8
django-cors-headers==4.4.0
django-extensions==3.2.3
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-spectacular==0.27.2
idna==3.7
inflection==0.5.1
isort==5.13.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.8.0
pylint==3.2.5
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
pyOpenSSL==24.2.1
python-dotenv==1.0.1
python3-openid==3.2.0
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.19.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
tomlkit==0.13.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
Werkzeug==3.0.3
